made the changes as per the review
